refactor(website): split ContactsVisual into a folder structure
ContactsVisual.tsx table shell + barrel, with ActorAvatar and CellValue
in components/, the four contact types split one-per-file into types/, and
COLUMNS / COMPANIES / HEADER_ICONS / ACTOR_SOURCE_ICONS in data/. The
shared table-grid styled and checkbox state stay in the shell.
