+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/data/actor-source-icons.ts
@@ -1,0 +1,11 @@
+import {
+  IconApi,
+  IconRobot,
+  IconSettingsAutomation,
+} from '@tabler/icons-react';
+
+export const ACTOR_SOURCE_ICONS = {
+  api: IconApi,
+  system: IconRobot,
+  workflow: IconSettingsAutomation,
+};
