+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/index.ts
@@ -1,0 +1,1 @@
+export { ContactsVisual } from './ContactsVisual';
