+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/types/contact-actor.ts
@@ -1,0 +1,10 @@
+type ContactTone = 'amber' | 'blue' | 'gray' | 'pink' | 'purple' | 'turquoise';
+
+type ActorSource = 'api' | 'system' | 'workflow';
+
+export type ContactActor = {
+  avatarUrl?: string;
+  name: string;
+  source?: ActorSource;
+  tone?: ContactTone;
+};
