+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/types/contact-column-id.ts
@@ -1,0 +1,9 @@
+export type ContactColumnId =
+  | 'accountOwner'
+  | 'address'
+  | 'arr'
+  | 'company'
+  | 'createdBy'
+  | 'icp'
+  | 'industry'
+  | 'url';
