+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/types/contact-column.ts
@@ -1,0 +1,8 @@
+import { type ContactColumnId } from './contact-column-id';
+
+export type ContactColumn = {
+  id: ContactColumnId;
+  isFirstColumn?: boolean;
+  label: string;
+  width: number;
+};
