+++ packages/twenty-website/src/sections/product-feature/ContactsVisual/types/contact-company.ts
@@ -1,0 +1,12 @@
+import { type ContactActor } from './contact-actor';
+
+export type ContactCompany = {
+  accountOwner: ContactActor;
+  address: string;
+  arr: string;
+  createdBy: ContactActor;
+  domain: string;
+  icp: boolean;
+  industry: string;
+  name: string;
+};
